# Populate pattern library and close out discovery work