--- a/src/pattern-library/pattern-generator.js
+++ b/src/pattern-library/pattern-generator.js
@@ -1,15 +1,27 @@
 const fs = require('fs');
 const path = require('path');
+const chalk = require('chalk');
 
 const pathIndex = process.argv.indexOf('-p');
 const nameIndex = process.argv.indexOf('-n');
 const titleIndex = process.argv.indexOf('-t');
+const skipMarkupIndex = process.argv.indexOf('-sm');
 const basePath = [process.cwd(), 'src', 'pattern-library', 'patterns'];
+
+// Set up the chalk warning and error states
+const warning = chalk.black.bgYellow;
+const error = chalk.black.bgRed;
+const success = chalk.black.bgGreen;
 
 if (pathIndex > 0 && nameIndex > 0) {
   const newPath = process.argv.slice(pathIndex + 1)[0];
   const name = process.argv.slice(nameIndex + 1)[0];
   const title = process.argv.slice(titleIndex + 1)[0];
+  const isVariant = newPath.includes('variants');
+
+  // This one is only used for variants
+  const skipMarkup =
+    process.argv.slice(skipMarkupIndex + 1)[0] === 'true' ? true : false;
 
   basePath.push(newPath);
 
@@ -22,17 +34,23 @@ if (pathIndex > 0 && nameIndex > 0) {
     fs.mkdirSync(path.join(...basePath), {recursive: true});
   }
 
-  fs.writeFileSync(path.join(...[...basePath, `${name}.njk`]), '');
+  // Create markup if not skipping (variants only)
+  if (isVariant && skipMarkup) {
+    console.log(warning('Markup skipped for this variant'));
+  } else {
+    fs.writeFileSync(path.join(...[...basePath, `${name}.njk`]), '');
+  }
+
   fs.writeFileSync(
     path.join(...[...basePath, `${name}.json`]),
     `{ "title": "${title || name}" }`,
   );
 
-  if (!newPath.includes('variants')) {
+  if (!isVariant) {
     fs.writeFileSync(path.join(...[...basePath, `${name}.md`]), '');
   }
 
-  console.log('Pattern created!');
+  console.log(success(`${isVariant ? 'Variant' : 'Pattern'} created!`));
 } else {
-  console.error('Name (-n) and/or Path (-p) not defined');
+  console.log(error('Name (-n) and/or Path (-p) not defined'));
 }