--- a/src/pattern-library/patterns/aside/aside.njk
+++ b/src/pattern-library/patterns/aside/aside.njk
@@ -1,0 +1,3 @@
+<div class="w-aside w-aside--{{ data.type if data.type else 'note' }}">
+  <p>{% if data.title %}<strong>{{ data.title }}</strong>: {% endif %}{{ data.content | safe }}</p>
+</div>