--- a/src/pattern-library/patterns/authors/variants/card-context/card-context.njk
+++ b/src/pattern-library/patterns/authors/variants/card-context/card-context.njk
@@ -1,0 +1,43 @@
+<div class="w-authors__card">
+  <div class="w-author__image--row">
+    {% for author in data.authors %}
+    <div class="w-author__image--row-item">
+      <a href="{{ author.url }}">
+        <img
+          alt="{{ author.name }}"
+          class="w-author__image w-author__image--small"
+          height="40"
+          src="https://web-dev.imgix.net{{
+            author.image
+          }}?auto=format&amp;fit=crop&amp;h=40&amp;w=40"
+          width="40"
+          loading="lazy"
+          decoding="async"
+          sizes="(min-width: 40px) 40px, calc(100vw - 48px)"
+          srcset="
+            https://web-dev.imgix.net{{ author.image }}?fit=crop&amp;h=40&amp;w=40&amp;auto=format&amp;dpr=1&amp;q=75,
+            https://web-dev.imgix.net{{ author.image }}?fit=crop&amp;h=40&amp;w=40&amp;auto=format&amp;dpr=2&amp;q=50 2x,
+            https://web-dev.imgix.net{{ author.image }}?fit=crop&amp;h=40&amp;w=40&amp;auto=format&amp;dpr=3&amp;q=35 3x,
+            https://web-dev.imgix.net{{ author.image }}?fit=crop&amp;h=40&amp;w=40&amp;auto=format&amp;dpr=4&amp;q=23 4x,
+            https://web-dev.imgix.net{{ author.image }}?fit=crop&amp;h=40&amp;w=40&amp;auto=format&amp;dpr=5&amp;q=20 5x
+          "
+        />
+      </a>
+    </div>
+    {% endfor %}
+  </div>
+  <div class="w-authors__card--holder">
+    <span class="w-author__name">
+      {% for author in data.authors %}
+      <!--prettier-ignore-->
+      <a href="{{ author.url }}" class="w-author__name-link">{{ author.name }}</a>
+      <!--prettier-ignore-->
+      {% if not loop.last %},{% endif %}
+      <!--prettier-ignore-->
+      {% endfor %}
+    </span>
+    <div class="w-author__published">
+      <time>{{ data.date }}</time>
+    </div>
+  </div>
+</div>