--- a/src/pattern-library/patterns/banners/banners.njk
+++ b/src/pattern-library/patterns/banners/banners.njk
@@ -1,0 +1,3 @@
+<div class="w-banner {{ data.modifier }}" role="banner">
+  {{ data.content | safe }}
+</div>