--- a/src/pattern-library/patterns/block-quotes/block-quotes.njk
+++ b/src/pattern-library/patterns/block-quotes/block-quotes.njk
@@ -1,0 +1,6 @@
+<blockquote>
+  <p>{{ data.quote }}</p>
+  {% if data.cite %}
+  <cite>{{ data.cite }}</cite>
+  {% endif %}
+</blockquote>