--- a/src/pattern-library/patterns/button/button.njk
+++ b/src/pattern-library/patterns/button/button.njk
@@ -1,1 +1,1 @@
-<button class="w-button">{{ data.text }}</button>
+<button class="w-button {{ data.modifier }}">{{ data.text }}</button>