--- a/src/pattern-library/patterns/button/button.njk
+++ b/src/pattern-library/patterns/button/button.njk
@@ -1,1 +1,1 @@
-<button class="w-button {{ data.modifier }}">{{ data.text }}</button>
+<button class="w-button {{ data.modifier }}"{% if data.icon %} data-icon="{{ data.icon }}"{% endif %}>{{ data.text }}</button>