--- a/src/pattern-library/patterns/button/variants/button-primary/button-primary.njk
+++ b/src/pattern-library/patterns/button/variants/button-primary/button-primary.njk
@@ -1,0 +1,1 @@
+<button class="w-button w-button--primary">{{ data.text }}</button>