--- a/src/pattern-library/patterns/button/variants/button-secondary/button-secondary.njk
+++ b/src/pattern-library/patterns/button/variants/button-secondary/button-secondary.njk
@@ -1,1 +1,0 @@
-<button class="w-button w-button--secondary">{{ data.text }}</button>