--- a/src/pattern-library/patterns/cards/variants/path/path.njk
+++ b/src/pattern-library/patterns/cards/variants/path/path.njk
@@ -1,0 +1,39 @@
+<a href="/reliable" class="w-card" role="listitem"
+  ><div class="w-path-card">
+    <div class="w-path-card__info">
+      <ul class="w-path-card__info-list">
+        <li
+          class="w-path-card__info-listitem w-path-card__info-listitem--category"
+        >
+          Collection
+        </li>
+        <li
+          class="w-path-card__info-listitem w-path-card__info-listitem--more-info"
+        >
+          23 resources
+        </li>
+        <li
+          class="w-path-card__info-listitem w-path-card__info-listitem--updated"
+        >
+          Updated <time></time>
+        </li>
+      </ul>
+    </div>
+    <div class="w-path-card__cover">
+      <img
+        class="w-path-card__cover-image"
+        src="https://web-dev.imgix.net/image/jxu1OdD7LKOGIDU7jURMpSH2lyK2/9K5gvnLRtD0yyiJM2r0c.svg"
+        alt=""
+        height="240"
+        width="100%"
+        loading="lazy"
+      />
+    </div>
+    <div class="w-path-card__desc">
+      <h2 class="w-path-card__headline">Network reliability</h2>
+      <p class="w-path-card__subhead">
+        See consistent, reliable performance regardless of network quality.
+      </p>
+    </div>
+  </div></a
+>