--- a/src/pattern-library/patterns/checkbox/checkbox.njk
+++ b/src/pattern-library/patterns/checkbox/checkbox.njk
@@ -1,0 +1,9 @@
+<div class="w-display--inline-flex">
+  <input
+    id="{{ data.id }}"
+    name="{{ data.name }}"
+    class="w-checkbox"
+    type="checkbox"
+  />
+  <label for="{{ data.id }}" class="w-ml--l">{{ data.label }}</label>
+</div>