--- a/src/pattern-library/patterns/compare/compare.njk
+++ b/src/pattern-library/patterns/compare/compare.njk
@@ -1,0 +1,14 @@
+{% for item in data.items %}
+<figure class="w-compare">
+  <p class="w-compare__label w-compare__label--{{ item.type }}">
+    {{ item.label }}
+  </p>
+  {{ item.content | md | safe }}
+
+  {% if item.caption %}
+  <figcaption class="w-compare__caption">
+    <p>{{ item.caption | md | safe }}</p>
+  </figcaption>
+  {% endif %}
+</figure>
+{% endfor %}