--- a/src/pattern-library/patterns/details/details.njk
+++ b/src/pattern-library/patterns/details/details.njk
@@ -1,0 +1,6 @@
+<details class="w-details" {{ data.state | safe }}>
+  <summary class="w-details__summary">
+    <h2 class="w-details__header">{{ data.summary }}</h2>
+  </summary>
+  {{ data.content | safe }}
+</details>