--- a/src/pattern-library/patterns/details/variants/extra-summary-content/extra-summary-content.njk
+++ b/src/pattern-library/patterns/details/variants/extra-summary-content/extra-summary-content.njk
@@ -1,0 +1,7 @@
+<details class="w-details">
+  <summary class="w-details__summary">
+    <h2 class="w-details__header">{{ data.summary }}</h2>
+    <p class="w-details__preview">{{ data.summaryPreview }}</p>
+  </summary>
+  {{ data.content | safe }}
+</details>