--- a/src/pattern-library/patterns/form-fields/form-fields.njk
+++ b/src/pattern-library/patterns/form-fields/form-fields.njk
@@ -1,0 +1,8 @@
+<label for="{{ data.id }}" class="w-visually-hidden">{{ data.label }}</label>
+<input
+  id="{{ data.id }}"
+  name="{{ data.name }}"
+  type="{{ data.type }}"
+  required
+  placeholder="{{ data.label }}"
+/>