--- a/src/pattern-library/patterns/form-fields/variants/radio/radio.njk
+++ b/src/pattern-library/patterns/form-fields/variants/radio/radio.njk
@@ -1,0 +1,13 @@
+{% for item in data.items %}
+<label for="{{ item.id }}">
+  {{ item.label }}
+  <!--prettier-ignore-->
+  <input
+    type="radio"
+    id="{{ item.id }}"
+    name="{{ data.group }}"
+    value="{{ item.value }}"
+    {% if item.checked %}checked{% endif %}
+  />
+</label>
+{% endfor %}