--- a/src/pattern-library/patterns/form-fields/variants/select/select.njk
+++ b/src/pattern-library/patterns/form-fields/variants/select/select.njk
@@ -1,0 +1,11 @@
+<label for="{{ data.id }}" class="w-visually-hidden">{{ data.label }}</label>
+<select
+  id="{{ data.id }}"
+  name="{{ data.name }}"
+  required=""
+  class="w-select--full-width"
+>
+  {% for item in data.items %}
+  <option value="{{ item.value }}">{{ item.text }}</option>
+  {% endfor %}
+</select>