--- a/src/pattern-library/patterns/headline/headline.njk
+++ b/src/pattern-library/patterns/headline/headline.njk
@@ -1,0 +1,1 @@
+<h2 class="w-headline{{ data.modifier }}">{{ data.title }}</h2>