--- a/src/pattern-library/patterns/icon-button/icon-button.njk
+++ b/src/pattern-library/patterns/icon-button/icon-button.njk
@@ -1,0 +1,3 @@
+<button class="w-button--icon {{ data.modifiers }}" data-icon="{{ data.icon }}">
+  <span class="w-tooltip" role="tooltip">{{ data.tooltip }}</span>
+</button>