--- a/src/pattern-library/patterns/icon-list/icon-list.njk
+++ b/src/pattern-library/patterns/icon-list/icon-list.njk
@@ -1,0 +1,7 @@
+<ul class="w-icon-list">
+  {% for item in data.items %}
+    <li class="w-icon-list__item w-icon-list__item--{{ data.icon }}">
+      <a href="{{ item.url }}" class="w-path-link">{{ item.text }}</a>
+    </li>
+  {% endfor %}
+</ul>