--- a/src/pattern-library/patterns/labels/labels.njk
+++ b/src/pattern-library/patterns/labels/labels.njk
@@ -1,0 +1,2 @@
+<p class="w-label">filename.js:</p>
+<pre><code>console.log('hello');</code></pre>