--- a/src/pattern-library/patterns/lists/lists.njk
+++ b/src/pattern-library/patterns/lists/lists.njk
@@ -1,0 +1,5 @@
+{% for item in data.items %}
+<ol>
+  <li>{{ item }}</li>
+</ol>
+{% endfor %}