--- a/src/pattern-library/patterns/lists/variants/unordered/unordered.njk
+++ b/src/pattern-library/patterns/lists/variants/unordered/unordered.njk
@@ -1,0 +1,5 @@
+{% for item in data.items %}
+<ul>
+  <li>{{ item }}</li>
+</ul>
+{% endfor %}