--- a/src/pattern-library/patterns/numbered-headers/numbered-headers.njk
+++ b/src/pattern-library/patterns/numbered-headers/numbered-headers.njk
@@ -1,0 +1,10 @@
+<section class="w-numbered-headers">
+  <div class="w-measure-steps">
+    {% for step in data.steps %}
+      <div class="w-measure-step">
+        <h3 class="w-numbered-header {{ data.direction }}">{{ step.title }}</h3>
+        <div>{{ step.summary }}</div>
+      </div>
+    {% endfor %}
+  </div>
+</section>