--- a/src/pattern-library/patterns/page-header/page-header.njk
+++ b/src/pattern-library/patterns/page-header/page-header.njk
@@ -1,0 +1,6 @@
+<header class="w-page-header">
+  <h1 class="w-page-header__headline">{{ data.title }}</h1>
+  {% if data.summary %}
+    <p class="w-page-header__copy">{{ data.summary }}</p>
+  {% endif %}
+</header>