--- a/src/pattern-library/patterns/stats/stats.njk
+++ b/src/pattern-library/patterns/stats/stats.njk
@@ -1,0 +1,13 @@
+<div class="w-stats">
+  {% for item in data.items %}
+  <div class="w-stat">
+    <p class="w-stat__figure">
+      {{ item.stat }}
+      {% if item.sub %}
+      <sub class="w-stat__sub">{{ item.sub }}</sub>
+      {% endif %}
+    </p>
+    <p class="w-stat__desc">{{ item.desc }}</p>
+  </div>
+  {% endfor %}
+</div>