--- a/src/pattern-library/patterns/toggle-switch/toggle-switch.njk
+++ b/src/pattern-library/patterns/toggle-switch/toggle-switch.njk
@@ -1,7 +1,0 @@
-<label class="toggle" for="{{ data.id }}">
-  <span class="toggle__label">{{ data.label }}</span>
-  <input type="checkbox" role="switch" class="toggle__element" id="{{ data.id }}" />
-  <div class="toggle__decor" aria-hidden="true">
-    <div class="toggle__thumb"></div>
-  </div>
-</label>