--- a/src/pattern-library/patterns/tooltips/tooltips.njk
+++ b/src/pattern-library/patterns/tooltips/tooltips.njk
@@ -1,0 +1,5 @@
+<button class="w-button--icon" data-icon="{{ data.icon }}">
+  <span class="w-tooltip {{ data.modifier }}" role="tooltip">{{
+    data.tooltip
+  }}</span>
+</button>