--- a/src/site/content/en/design-system/patterns.njk
+++ b/src/site/content/en/design-system/patterns.njk
@@ -1,7 +1,7 @@
 ---
 title: 'Pattern'
 pagination:
-  data: patterns.items
+  data: design.patterns.items
   size: 1
   addAllPagesToCollections: true
   alias: item