A big ol' PR full of pattern population and getting a good reference of the site as it currently stands. 

Here's a useful screencast: https://www.loom.com/share/e80584f512a14e62814b029938ef444c

Fixes #5376