@robdodson 2021-06-26
Taking a look at this now. It seems like there's possibly a bug somewhere that is causing the build times to go up.

On this branch a build takes around 70 seconds:
```
Eleventy is building, please wait…
Wrote 1102 files in 70.00 seconds (63.5ms each, v0.11.1)
```

But on the current main a build takes around 19 seconds.
```
Eleventy is building, please wait…
Wrote 1018 files in 19.47 seconds (19.1ms each, v0.11.1)
```

I'll read through the code to see if anything jumps out at me.