@robdodson 2021-06-26
Some benchmarks in case they help:

**This branch**
```
➜  web.dev git:(populate-pattern-library) ✗ DEBUG=Eleventy:Benchmark* npx @11ty/eleventy

Eleventy is building, please wait…
  Eleventy:Benchmark Benchmark (Configuration): "slug" Nunjucks Filter took 3ms (0.0%, called 294×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "url" Nunjucks Filter took 31ms (0.0%, called 858×, 0.0ms each) +1ms
  Eleventy:Benchmark Benchmark (Configuration): "highlight" Nunjucks Paired Shortcode took 55ms (0.1%, called 132×, 0.4ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "absoluteUrl" Nunjucks Filter took 10ms (0.0%, called 858×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "htmlToAbsoluteUrls" Nunjucks Async Filter took 1449ms (2.1%, called 858×, 1.7ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "rssDate" Nunjucks Filter took 3ms (0.0%, called 861×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "toc" Nunjucks Filter took 2220ms (3.2%, called 610×, 3.6ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "i18n" Nunjucks Filter took 29ms (0.0%, called 25323×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "findTags" Nunjucks Filter took 2ms (0.0%, called 610×, 0.0ms each) +1ms
  Eleventy:Benchmark Benchmark (Configuration): "expandAuthors" Nunjucks Filter took 2ms (0.0%, called 635×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "githubLink" Nunjucks Filter took 3ms (0.0%, called 610×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "gitlocalizeLink" Nunjucks Filter took 1ms (0.0%, called 967×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "htmlDateString" Nunjucks Filter took 39ms (0.1%, called 691×, 0.1ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "md" Nunjucks Filter took 18ms (0.0%, called 1345×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "navigation" Nunjucks Filter took 6ms (0.0%, called 27×, 0.2ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "pagedNavigation" Nunjucks Filter took 2ms (0.0%, called 269×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "postsLighthouseJson" Nunjucks Filter took 1ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "prettyDate" Nunjucks Filter took 57ms (0.1%, called 1607×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "removeDrafts" Nunjucks Filter took 1ms (0.0%, called 22×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "slugify" Nunjucks Filter took 31ms (0.0%, called 1220×, 0.0ms each) +1ms
  Eleventy:Benchmark Benchmark (Configuration): "getPaths" Nunjucks Filter took 1ms (0.0%, called 385×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "courseToc" Nunjucks Filter took 5ms (0.0%, called 25×, 0.2ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "updateSvgForInclude" Nunjucks Filter took 16ms (0.0%, called 2896×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "minifyJs" Nunjucks Filter took 4ms (0.0%, called 2110×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "cspHash" Nunjucks Filter took 1ms (0.0%, called 2110×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "ArticleNavigation" Nunjucks Shortcode took 13ms (0.0%, called 628×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Aside" Nunjucks Paired Shortcode took 7ms (0.0%, called 1225×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Assessment" Nunjucks Shortcode took 122ms (0.2%, called 39×, 3.1ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Author" Nunjucks Shortcode took 133ms (0.2%, called 637×, 0.2ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "AuthorCard" Nunjucks Shortcode took 23ms (0.0%, called 160×, 0.1ms each) +1ms
  Eleventy:Benchmark Benchmark (Configuration): "Banner" Nunjucks Paired Shortcode took 39ms (0.1%, called 950×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Breadcrumbs" Nunjucks Shortcode took 6ms (0.0%, called 286×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "CodelabsCallout" Nunjucks Shortcode took 2ms (0.0%, called 42×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Codepen" Nunjucks Shortcode took 3ms (0.0%, called 239×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Compare" Nunjucks Paired Shortcode took 1ms (0.0%, called 220×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "DetailsSummary" Nunjucks Paired Shortcode took 2ms (0.0%, called 64×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Glitch" Nunjucks Shortcode took 5ms (0.0%, called 133×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Hero" Nunjucks Shortcode took 38ms (0.1%, called 277×, 0.1ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Img" Nunjucks Shortcode took 260ms (0.4%, called 1895×, 0.1ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Instruction" Nunjucks Shortcode took 2ms (0.0%, called 213×, 0.0ms each) +1ms
  Eleventy:Benchmark Benchmark (Configuration): "Label" Nunjucks Paired Shortcode took 1ms (0.0%, called 77×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Meta" Nunjucks Shortcode took 181ms (0.3%, called 1030×, 0.2ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "PathCard" Nunjucks Shortcode took 3ms (0.0%, called 38×, 0.1ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "PostCard" Nunjucks Shortcode took 561ms (0.8%, called 1872×, 0.3ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "SignPosts" Nunjucks Shortcode took 9ms (0.0%, called 610×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Video" Nunjucks Shortcode took 7ms (0.0%, called 107×, 0.1ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "YouTube" Nunjucks Shortcode took 1ms (0.0%, called 73×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "EventTable" Nunjucks Shortcode took 17ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Aggregate): Searching the file system took 51ms (0.1%, called 2×, 25.5ms each) +0ms
  Eleventy:Benchmark Benchmark (Aggregate): Data File took 789ms (1.1%, called 2964×, 0.3ms each) +1ms
  Eleventy:Benchmark Benchmark (Aggregate): Template Read took 36284ms (52.4%, called 1209×, 30.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Aggregate): Template Compile took 43ms (0.1%, called 9197×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Aggregate): Template Render took 4029ms (5.8%, called 7048×, 0.6ms each) +0ms
  Eleventy:Benchmark Benchmark (Aggregate): Template Write took 6456ms (9.3%, called 1102×, 5.9ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/_data/countries.js` took 1ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/_data/event.js` took 1ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/_data/postHost.js` took 2ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/_data/postToPaths.js` took 1ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/_data/process.js` took 1ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/_data/design/pages.js` took 1ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/_data/design/patterns.js` took 25ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/_data/paths/index.js` took 1ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/content.11tydata.js` took 265ms (0.4%, called 1113×, 0.2ms each) +1ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/en.11tydata.js` took 323ms (0.5%, called 1101×, 0.3ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/es/es.11tydata.js` took 2ms (0.0%, called 5×, 0.5ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/pl/pl.11tydata.js` took 2ms (0.0%, called 4×, 0.5ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/design-system/design-system.11tydata.json` took 16ms (0.0%, called 107×, 0.2ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/index.11tydata.js` took 1ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/es/index.11tydata.js` took 1ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/pl/index.11tydata.js` took 1ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/blog/blog.11tydata.js` took 44ms (0.1%, called 284×, 0.2ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/learn/learn.11tydata.js` took 8ms (0.0%, called 26×, 0.3ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/newsletter/newsletter.11tydata.js` took 1ms (0.0%, called 7×, 0.2ms each) +1ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/authors/feed.11tydata.js` took 3ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/authors/index.11tydata.js` took 2ms (0.0%, called 8×, 0.2ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/authors/individual.11tydata.js` took 41ms (0.1%, called 167×, 0.2ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/blog/index.11tydata.js` took 4ms (0.0%, called 13×, 0.3ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/newsletter/index.11tydata.js` took 1ms (0.0%, called 2×, 0.6ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/tags/feed.11tydata.js` took 1ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/tags/individual.11tydata.js` took 20ms (0.0%, called 81×, 0.2ms each) +1ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/tags/index.11tydata.js` took 1ms (0.0%, called 4×, 0.3ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/learn/css/css.11tydata.js` took 6ms (0.0%, called 25×, 0.2ms each) +0ms
Wrote 1102 files in 68.44 seconds (62.1ms each, v0.11.1)


Start time: (2021-06-26 00:28:04 UTC) [treshold=1%]
#  module                                                 time  %
1  /Users/robdodson/Develope...eventy.js (.eleventy.js)  760ms  ▇ 1%
2  ./Config (node_modules/@1.../eleventy/src/Config.js)  844ms  ▇ 1%
3  ./TemplateEngineManager (...emplateEngineManager.js)  845ms  ▇ 1%
4  ./EleventyExtensionMap (n...EleventyExtensionMap.js)  845ms  ▇ 1%
5  ./TemplateRender (node_mo...y/src/TemplateRender.js)  846ms  ▇ 1%
6  ./TemplateData (node_modu...nty/src/TemplateData.js)  967ms  ▇ 1%
7  ./src/Eleventy (node_modu...leventy/src/Eleventy.js)     1s  ▇ 1%
Total require(): 17289
Total time: 1m 9.5s
➜  web.dev git:(populate-pattern-library) 
```

**main branch**
```
Eleventy is building, please wait…
  Eleventy:Benchmark Benchmark (Configuration): "slug" Nunjucks Filter took 3ms (0.0%, called 294×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "url" Nunjucks Filter took 5ms (0.0%, called 72×, 0.1ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "highlight" Nunjucks Paired Shortcode took 7ms (0.0%, called 6×, 1.1ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "absoluteUrl" Nunjucks Filter took 1ms (0.0%, called 72×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "htmlToAbsoluteUrls" Nunjucks Async Filter took 180ms (0.8%, called 72×, 2.5ms each) +0ms
Benchmark (Configuration): "toc" Nunjucks Filter took 2235ms (10.2%, called 617×, 3.6ms each)
  Eleventy:Benchmark Benchmark (Configuration): "toc" Nunjucks Filter took 2235ms (10.2%, called 617×, 3.6ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "i18n" Nunjucks Filter took 31ms (0.1%, called 26294×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "expandAuthors" Nunjucks Filter took 2ms (0.0%, called 642×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "githubLink" Nunjucks Filter took 2ms (0.0%, called 617×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "htmlDateString" Nunjucks Filter took 42ms (0.2%, called 699×, 0.1ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "md" Nunjucks Filter took 18ms (0.1%, called 1352×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "navigation" Nunjucks Filter took 6ms (0.0%, called 27×, 0.2ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "pagedNavigation" Nunjucks Filter took 2ms (0.0%, called 274×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "postsLighthouseJson" Nunjucks Filter took 1ms (0.0%) +1ms
  Eleventy:Benchmark Benchmark (Configuration): "prettyDate" Nunjucks Filter took 55ms (0.3%, called 1623×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "removeDrafts" Nunjucks Filter took 1ms (0.0%, called 22×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "slugify" Nunjucks Filter took 20ms (0.1%, called 1234×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "getPaths" Nunjucks Filter took 1ms (0.0%, called 386×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "courseToc" Nunjucks Filter took 6ms (0.0%, called 25×, 0.2ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "updateSvgForInclude" Nunjucks Filter took 13ms (0.1%, called 2821×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "minifyJs" Nunjucks Filter took 3ms (0.0%, called 2068×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "cspHash" Nunjucks Filter took 1ms (0.0%, called 2068×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "ArticleNavigation" Nunjucks Shortcode took 19ms (0.1%, called 635×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Aside" Nunjucks Paired Shortcode took 8ms (0.0%, called 1240×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Assessment" Nunjucks Shortcode took 133ms (0.6%, called 39×, 3.4ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Author" Nunjucks Shortcode took 102ms (0.5%, called 649×, 0.2ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "AuthorsDate" Nunjucks Shortcode took 327ms (1.5%, called 1836×, 0.2ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Banner" Nunjucks Paired Shortcode took 1ms (0.0%, called 30×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Breadcrumbs" Nunjucks Shortcode took 6ms (0.0%, called 292×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "CodelabsCallout" Nunjucks Shortcode took 2ms (0.0%, called 42×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Codepen" Nunjucks Shortcode took 3ms (0.0%, called 239×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Compare" Nunjucks Paired Shortcode took 1ms (0.0%, called 220×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "DetailsSummary" Nunjucks Paired Shortcode took 2ms (0.0%, called 69×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Glitch" Nunjucks Shortcode took 9ms (0.0%, called 134×, 0.1ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Hero" Nunjucks Shortcode took 55ms (0.2%, called 283×, 0.2ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Img" Nunjucks Shortcode took 377ms (1.7%, called 3337×, 0.1ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Instruction" Nunjucks Shortcode took 2ms (0.0%, called 213×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Label" Nunjucks Paired Shortcode took 1ms (0.0%, called 77×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Meta" Nunjucks Shortcode took 172ms (0.8%, called 1009×, 0.2ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "PathCard" Nunjucks Shortcode took 3ms (0.0%, called 38×, 0.1ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "SignPosts" Nunjucks Shortcode took 10ms (0.0%, called 617×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "Video" Nunjucks Shortcode took 7ms (0.0%, called 107×, 0.1ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "YouTube" Nunjucks Shortcode took 1ms (0.0%, called 73×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Configuration): "EventTable" Nunjucks Shortcode took 16ms (0.1%) +0ms
  Eleventy:Benchmark Benchmark (Aggregate): Searching the file system took 49ms (0.2%, called 2×, 24.4ms each) +0ms
  Eleventy:Benchmark Benchmark (Aggregate): Data File took 683ms (3.1%, called 2708×, 0.3ms each) +1ms
  Eleventy:Benchmark Benchmark (Aggregate): Template Read took 6515ms (29.7%, called 1125×, 5.8ms each) +0ms
  Eleventy:Benchmark Benchmark (Aggregate): Template Compile took 44ms (0.2%, called 8633×, 0.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Aggregate): Template Render took 4521ms (20.6%, called 6464×, 0.7ms each) +0ms
  Eleventy:Benchmark Benchmark (Aggregate): Template Write took 4098ms (18.7%, called 1018×, 4.0ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/_data/countries.js` took 1ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/_data/event.js` took 1ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/_data/patterns.js` took 21ms (0.1%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/_data/podcasts.js` took 1ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/_data/postHost.js` took 2ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/_data/postToPaths.js` took 1ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/_data/process.js` took 1ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/content.11tydata.js` took 214ms (1.0%, called 1028×, 0.2ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/en.11tydata.js` took 281ms (1.3%, called 1016×, 0.3ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/es/es.11tydata.js` took 2ms (0.0%, called 5×, 0.4ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/pl/pl.11tydata.js` took 2ms (0.0%, called 4×, 0.5ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/design-system/design-system.11tydata.json` took 2ms (0.0%, called 10×, 0.2ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/index.11tydata.js` took 1ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/es/index.11tydata.js` took 1ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/pl/index.11tydata.js` took 1ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/blog/blog.11tydata.js` took 50ms (0.2%, called 291×, 0.2ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/learn/learn.11tydata.js` took 7ms (0.0%, called 26×, 0.3ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/newsletter/newsletter.11tydata.js` took 1ms (0.0%, called 7×, 0.2ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/authors/feed.11tydata.js` took 3ms (0.0%) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/authors/index.11tydata.js` took 2ms (0.0%, called 8×, 0.2ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/authors/individual.11tydata.js` took 43ms (0.2%, called 171×, 0.3ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/blog/index.11tydata.js` took 4ms (0.0%, called 14×, 0.3ms each) +0ms
  Eleventy:Benchmark Benchmark (Data): `./src/site/content/en/newsletter/index.11tydata.js` took 1ms (0.0%, called 2×, 0.7ms each) +0ms
…